refactor admin products to prisma
